% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{fricplot}
\alias{fricplot}
\title{Plot friction-estimation curves with uncertainty}
\usage{
fricplot(dT = 1, Angs, Sib., error.thresh = 100)
}
\arguments{
\item{dT}{numeric; the uncertainty in the lockup angle(s)}

\item{Angs}{numeric; lockup angle(s)}

\item{Sib.}{\code{"sibson.fric"} object}

\item{error.thresh}{numeric; the threshold of error estimates (in percent) to clip figure to}
}
\value{
A data.frame of the values shown (or not shown, depending on \code{thresh}), invisibly
}
\examples{
fricplot()
}
\seealso{
\code{\link{Sibson.fric}} and \code{\link{Mu}}
}

