% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friction.R
\name{optplot}
\alias{optplot}
\title{Plot optimal angles and stress ratios for fault failure}
\usage{
optplot(x, ...)
}
\arguments{
\item{x}{object to plot}

\item{...}{additional paramters sent to the main plotting function(s)}
}
\description{
Plot optimal angles and stress ratios for fault failure
}
\details{
Plots the variation in the optimum reactivation angle and minimum
(positive) stress ratio for reactivation as a function of friction.
Note that the stress ratio if calculated with \code{\link{StressRatio}}.
}
\examples{
optplot() # Reproduces Sibson (1985; Figure 2) by default
}
\seealso{
\code{\link{Sibson.fric}}
}

