% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{Utilities}
\alias{Utilities}
\alias{lower.saw}
\alias{upper.saw}
\alias{cotan}
\title{Utility functions}
\usage{
lower.saw(x, lim = min(x, na.rm = TRUE))

upper.saw(x, lim = max(x, na.rm = TRUE))

cotan(rad)
}
\arguments{
\item{x}{vector of values}

\item{lim}{numeric; the limit for sawtoothin'}

\item{rad}{numeric; angle in radians}
}
\description{
Various utility functions
}
