% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friction.R
\name{Sibson.fric}
\alias{Sibson.fric}
\title{Calculate failure criterion for a given set of frictional coefficients}
\usage{
Sibson.fric(mu, theta, thresh, verbose = TRUE)
}
\arguments{
\item{mu}{numeric; the static coefficient(s) of friction}

\item{theta}{numeric; optional; the orientation(s) of the fault with respect to the principal compressive stress}

\item{thresh}{numeric; an arbitrary clipping level for the stress ratio}

\item{verbose}{logical; should the optimal parameters be printed?}
}
\value{
An object with class \code{"sibson.fric"}
}
\description{
Calculate failure criterion for a given set of frictional coefficients
}
\examples{
# Use default values (Byerlee's law)
S <- Sibson.fric()
print(str(S))

# Use your own values
S <- Sibson.fric(0.5)
# and multiple values
S <- Sibson.fric(c(0.15,0.6))

# Plot the result
plot(S)
}
\seealso{
\code{\link{StressRatio}} and \code{\link{Stability}}; \code{\link{plot.sibson.fric}}
}
