% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{Mu}
\alias{.Mu.opt}
\alias{Mu}
\title{Static friction coefficient for a fault orientation}
\usage{
Mu(theta, ...)

.Mu.opt(theta, delta.theta = 0, in.deg = TRUE,
  precision = getOption("digits"))
}
\arguments{
\item{theta}{numeric; optional; the orientation(s) of the fault with respect to the principal compressive stress}

\item{delta.theta}{numeric; the uncertainty in theta, used to calculate uncertainty in the frictional coefficient}

\item{in.deg}{logical; is \code{theta} given in degrees?}

\item{precision}{integer; the numerical precision used to \code{\link{round}} the frictional coefficient}
}
\seealso{
\code{\link{Sibson.fric}}
}

