% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stability.R
\name{optplot}
\alias{optplot}
\title{Plot optimal angles and stress ratios for fault failure}
\usage{
optplot(x, ...)
}
\arguments{
\item{x}{object to plot}

\item{...}{additional paramters sent to the main plotting function(s)}
}
\examples{
optplot()
}
\seealso{
\code{\link{Sibson.fric}}
}

