% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sibson-package.R
\docType{package}
\name{sibson}
\alias{sibson}
\alias{sibson-package}
\title{Fault reactivation}
\description{
The functions in this package can be used to calculate various properties
of a fault in idealized rock, including frictional coefficients, stability angles, and
stress ratio, etc.
}
\references{
Sibson, R. H. (1985),
A note on fault reactivation,
\emph{Journal of Structural Geology}, \strong{7}(6),
751-754
\url{http://dx.doi.org/10.1016/0191-8141(85)90150-6}
}
\seealso{
\code{\link{Sibson.fric}}
}
\author{
A.J. Barbour
}
