% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/friction.R
\name{StressRatio}
\alias{StressRatio}
\title{Principal stress ratio}
\usage{
StressRatio(mu, theta, thresh = NULL, in.deg = TRUE)
}
\arguments{
\item{mu}{numeric; the static coefficient(s) of friction}

\item{theta}{numeric; optional; the orientation(s) of the fault with respect to the principal compressive stress}

\item{thresh}{numeric; an arbitrary clipping level for the stress ratio}

\item{in.deg}{logical; is \code{theta} given in degrees?}
}
\description{
Calculate the ratio of principal \emph{effective} stresses
}
\seealso{
\code{\link{Sibson.fric}}
}

